move training signals function to the global registrar so it can update even when components are dismounted
