+++ src/components/Resource/Forms/Training/TrainingCommon.ts
@@ -1,0 +1,77 @@
+import { useQueryClient } from '@tanstack/react-query';
+import { getQueryKey } from '@trpc/react-query';
+import produce from 'immer';
+import Router from 'next/router';
+import { useCallback } from 'react';
+import { useSignalConnection } from '~/components/Signals/SignalsProvider';
+import { SignalMessages } from '~/server/common/enums';
+import { TrainingUpdateSignalSchema } from '~/server/schema/signals.schema';
+import { MyTrainingModelGetAll } from '~/types/router';
+import { trpc } from '~/utils/trpc';
+
+export const basePath = '/models/train';
+export const maxSteps = 3;
+
+// these could use the current route to determine?
+export const goNext = (modelId: number | undefined, step: number) => {
+  if (modelId && step < maxSteps)
+    Router.replace(`${basePath}?modelId=${modelId}&step=${step + 1}`, undefined, {
+      shallow: true,
+      scroll: true,
+    });
+};
+export const goBack = (modelId: number | undefined, step: number) => {
+  if (modelId && step > 1)
+    Router.replace(`${basePath}?modelId=${modelId}&step=${step - 1}`, undefined, {
+      shallow: true,
+      scroll: true,
+    });
+};
+
+export const useTrainingSignals = () => {
+  const queryClient = useQueryClient();
+  const queryUtils = trpc.useContext();
+
+  const onUpdate = useCallback(
+    (updated: TrainingUpdateSignalSchema) => {
+      // Update training model view
+      const queryKey = getQueryKey(trpc.model.getMyTrainingModels);
+      queryClient.setQueriesData(
+        { queryKey, exact: false },
+        produce((old: MyTrainingModelGetAll | undefined) => {
+          const model = old?.items?.find((x) => x.id == updated.modelId);
+          const mv = model?.modelVersions[0];
+          if (mv) {
+            mv.trainingStatus = updated.status;
+            const mFile = mv.files[0];
+            if (mFile) {
+              mFile.metadata = updated.fileMetadata;
+            }
+          }
+        })
+      );
+
+      // Update select file page
+      queryUtils.model.getById.setData(
+        { id: updated.modelId },
+        produce((old) => {
+          if (!old) return old;
+          const mv = old.modelVersions[0];
+          if (mv) {
+            mv.trainingStatus = updated.status;
+            const mFile = mv.files.find((f) => f.type === 'Training Data');
+            if (mFile) {
+              // TODO [bw] why is this complaining about null in ModelFileFormat?
+              // @ts-ignore
+              mFile.metadata = updated.fileMetadata;
+            }
+          }
+        })
+      );
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [queryClient]
+  );
+
+  useSignalConnection(SignalMessages.TrainingUpdate, onUpdate);
+};
