+++ src/components/Signals/SignalsRegistrar.tsx
@@ -1,7 +1,9 @@
 import { useImageGenStatusUpdate } from '~/components/ImageGeneration/utils/generationRequestHooks';
+import { useTrainingSignals } from '~/components/Resource/Forms/Training/TrainingCommon';
 
 export function SignalsRegistrar() {
   useImageGenStatusUpdate();
+  useTrainingSignals();
 
   return null;
 }
